TR3-ASTRA/TR3-ASTRA/requirement.txt
scapy
prettytable
mac-vendor-lookup
progressbar2
termcolor
pyfiglet
requests
dnspython
whois
